Fixed error in type coercion of Variadic.
